reorg xmlenc1 tests into themed subtests
