+++ xmlenc1/decrypt_ns_test.go
@@ -9,268 +9,270 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// D-ENC-001: decrypting TypeContent plaintext that uses a namespace prefix
-// declared on an ANCESTOR of EncryptedData must succeed. The temporary parse
-// wrapper has to carry the in-scope namespace declarations so prefixes in the
-// decrypted fragment resolve.
-func TestDecryptContent_InScopeAncestorNamespace(t *testing.T) {
-	sessionKey := make([]byte, 32)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+func TestDecryptNamespace(t *testing.T) {
+	// D-ENC-001: decrypting TypeContent plaintext that uses a namespace prefix
+	// declared on an ANCESTOR of EncryptedData must succeed. The temporary parse
+	// wrapper has to carry the in-scope namespace declarations so prefixes in the
+	// decrypted fragment resolve.
+	t.Run("content in-scope ancestor namespace", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
 
-	algorithm := xmlenc1.AES256GCM
+		algorithm := xmlenc1.AES256GCM
 
-	// The decrypted content uses the saml: prefix, which is NOT declared in
-	// the fragment itself — only on an ancestor of EncryptedData.
-	plaintext := []byte(`<saml:NameID>user@example.com</saml:NameID>`)
-	cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
-	require.NoError(t, err)
+		// The decrypted content uses the saml: prefix, which is NOT declared in
+		// the fragment itself — only on an ancestor of EncryptedData.
+		plaintext := []byte(`<saml:NameID>user@example.com</saml:NameID>`)
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
+		require.NoError(t, err)
 
-	// Build a document where saml: is declared on the grandparent of
-	// EncryptedData, then splice the EncryptedData element in as content.
-	doc := mustParseXML(t,
-		`<saml:Subject xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion"><wrapper/></saml:Subject>`)
+		// Build a document where saml: is declared on the grandparent of
+		// EncryptedData, then splice the EncryptedData element in as content.
+		doc := mustParseXML(t,
+			`<saml:Subject xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion"><wrapper/></saml:Subject>`)
 
-	ed := &xmlenc1.EncryptedData{
-		Type:             xmlenc1.TypeContent,
-		EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
-		CipherValue:      cipher,
-	}
-	edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
-	require.NoError(t, err)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeContent,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	// Attach EncryptedData under <wrapper>, so saml: is in-scope via the
-	// ancestor chain (wrapper -> saml:Subject).
-	wrapper := doc.DocumentElement().FirstChild().(*helium.Element)
-	require.NoError(t, wrapper.AddChild(edElem))
+		// Attach EncryptedData under <wrapper>, so saml: is in-scope via the
+		// ancestor chain (wrapper -> saml:Subject).
+		wrapper := doc.DocumentElement().FirstChild().(*helium.Element)
+		require.NoError(t, wrapper.AddChild(edElem))
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "NameID", elem.LocalName())
-	require.Equal(t, "urn:oasis:names:tc:SAML:2.0:assertion", elem.URI(),
-		"saml: prefix must resolve to the ancestor-declared namespace URI")
-}
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "NameID", elem.LocalName())
+		require.Equal(t, "urn:oasis:names:tc:SAML:2.0:assertion", elem.URI(),
+			"saml: prefix must resolve to the ancestor-declared namespace URI")
+	})
 
-// D-ENC-002: decrypting TypeContent plaintext whose unprefixed elements must
-// resolve in the PARENT's default namespace — NOT in the XML-Encryption
-// namespace that EncryptedData itself declares. The decrypted content replaces
-// EncryptedData, so it must be parsed in the in-scope-namespace context of
-// EncryptedData's parent, not EncryptedData's own declarations.
-func TestDecryptContent_ParentDefaultNamespace(t *testing.T) {
-	sessionKey := make([]byte, 32)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+	// D-ENC-002: decrypting TypeContent plaintext whose unprefixed elements must
+	// resolve in the PARENT's default namespace — NOT in the XML-Encryption
+	// namespace that EncryptedData itself declares. The decrypted content replaces
+	// EncryptedData, so it must be parsed in the in-scope-namespace context of
+	// EncryptedData's parent, not EncryptedData's own declarations.
+	t.Run("content parent default namespace", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
 
-	algorithm := xmlenc1.AES256GCM
+		algorithm := xmlenc1.AES256GCM
 
-	// The decrypted content is an unprefixed element. At its replacement
-	// position (a child of <wrapper> in the app namespace) it must inherit
-	// the parent's default namespace.
-	plaintext := []byte(`<child/>`)
-	cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
-	require.NoError(t, err)
+		// The decrypted content is an unprefixed element. At its replacement
+		// position (a child of <wrapper> in the app namespace) it must inherit
+		// the parent's default namespace.
+		plaintext := []byte(`<child/>`)
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
+		require.NoError(t, err)
 
-	// The parent declares a default application namespace that differs from
-	// the XML-Encryption namespace EncryptedData declares for itself.
-	const appNS = "http://example.com/app"
-	doc := mustParseXML(t,
-		`<root xmlns="`+appNS+`"><wrapper/></root>`)
+		// The parent declares a default application namespace that differs from
+		// the XML-Encryption namespace EncryptedData declares for itself.
+		const appNS = "http://example.com/app"
+		doc := mustParseXML(t,
+			`<root xmlns="`+appNS+`"><wrapper/></root>`)
 
-	ed := &xmlenc1.EncryptedData{
-		Type:             xmlenc1.TypeContent,
-		EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
-		CipherValue:      cipher,
-	}
-	edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
-	require.NoError(t, err)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeContent,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	// EncryptedData declares its OWN default namespace (the XML-Encryption
-	// namespace). Parsing the decrypted content with EncryptedData's own
-	// declarations would wrongly resolve <child/> into XMLENC; it must use
-	// the parent's in-scope default namespace instead.
-	require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
+		// EncryptedData declares its OWN default namespace (the XML-Encryption
+		// namespace). Parsing the decrypted content with EncryptedData's own
+		// declarations would wrongly resolve <child/> into XMLENC; it must use
+		// the parent's in-scope default namespace instead.
+		require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
 
-	wrapper := doc.DocumentElement().FirstChild().(*helium.Element)
-	require.NoError(t, wrapper.AddChild(edElem))
+		wrapper := doc.DocumentElement().FirstChild().(*helium.Element)
+		require.NoError(t, wrapper.AddChild(edElem))
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "child", elem.LocalName())
-	require.Equal(t, appNS, elem.URI(),
-		"unprefixed element must resolve in the parent's default namespace, not XMLENC")
-}
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "child", elem.LocalName())
+		require.Equal(t, appNS, elem.URI(),
+			"unprefixed element must resolve in the parent's default namespace, not XMLENC")
+	})
 
-// D-ENC-004: decrypting TypeContent for a DETACHED EncryptedData (no parent)
-// whose own default xmlns is the XML-Encryption namespace must NOT let that
-// declaration shadow the decrypted content. Unprefixed plaintext <child/> must
-// resolve to NO namespace, not the XML-Encryption namespace, because a detached
-// element has no replacement context and its own xmlns decls are irrelevant to
-// the fragment's namespace resolution.
-func TestDecryptContent_DetachedNoShadowing(t *testing.T) {
-	sessionKey := make([]byte, 32)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+	// D-ENC-004: decrypting TypeContent for a DETACHED EncryptedData (no parent)
+	// whose own default xmlns is the XML-Encryption namespace must NOT let that
+	// declaration shadow the decrypted content. Unprefixed plaintext <child/> must
+	// resolve to NO namespace, not the XML-Encryption namespace, because a detached
+	// element has no replacement context and its own xmlns decls are irrelevant to
+	// the fragment's namespace resolution.
+	t.Run("content detached no shadowing", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
 
-	algorithm := xmlenc1.AES256GCM
+		algorithm := xmlenc1.AES256GCM
 
-	plaintext := []byte(`<child/>`)
-	cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
-	require.NoError(t, err)
+		plaintext := []byte(`<child/>`)
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
+		require.NoError(t, err)
 
-	// Owning document exists but EncryptedData is never attached to it.
-	doc := mustParseXML(t, `<root/>`)
+		// Owning document exists but EncryptedData is never attached to it.
+		doc := mustParseXML(t, `<root/>`)
 
-	ed := &xmlenc1.EncryptedData{
-		Type:             xmlenc1.TypeContent,
-		EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
-		CipherValue:      cipher,
-	}
-	edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
-	require.NoError(t, err)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeContent,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	// EncryptedData declares its OWN default namespace (the XML-Encryption
-	// namespace) and remains DETACHED (no parent). Using EncryptedData itself
-	// as the parse context would wrongly resolve <child/> into XMLENC; with no
-	// replacement parent the content must resolve in no namespace.
-	require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
-	require.Nil(t, edElem.Parent(), "EncryptedData must be detached for this test")
+		// EncryptedData declares its OWN default namespace (the XML-Encryption
+		// namespace) and remains DETACHED (no parent). Using EncryptedData itself
+		// as the parse context would wrongly resolve <child/> into XMLENC; with no
+		// replacement parent the content must resolve in no namespace.
+		require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
+		require.Nil(t, edElem.Parent(), "EncryptedData must be detached for this test")
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "child", elem.LocalName())
-	require.Equal(t, "", elem.URI(),
-		"detached EncryptedData's own xmlns must not shadow the decrypted content")
-}
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "child", elem.LocalName())
+		require.Equal(t, "", elem.URI(),
+			"detached EncryptedData's own xmlns must not shadow the decrypted content")
+	})
 
-// D-ENC-005: encrypting then decrypting an ELEMENT whose tag uses a namespace
-// prefix declared only on an ANCESTOR must round-trip. The serialized plaintext
-// (<saml:NameID>...</saml:NameID>) carries no xmlns:saml declaration, so parsing
-// it as a standalone document fails on the unbound prefix. The decryption must
-// parse it in the in-scope-namespace context of EncryptedData's parent instead.
-func TestEncryptDecryptElement_AncestorNamespace(t *testing.T) {
-	const samlNS = "urn:oasis:names:tc:SAML:2.0:assertion"
-	doc := mustParseXML(t,
-		`<root xmlns:saml="`+samlNS+`"><saml:NameID>user@example.com</saml:NameID></root>`)
+	// D-ENC-005: encrypting then decrypting an ELEMENT whose tag uses a namespace
+	// prefix declared only on an ANCESTOR must round-trip. The serialized plaintext
+	// (<saml:NameID>...</saml:NameID>) carries no xmlns:saml declaration, so parsing
+	// it as a standalone document fails on the unbound prefix. The decryption must
+	// parse it in the in-scope-namespace context of EncryptedData's parent instead.
+	t.Run("element ancestor namespace round-trip", func(t *testing.T) {
+		const samlNS = "urn:oasis:names:tc:SAML:2.0:assertion"
+		doc := mustParseXML(t,
+			`<root xmlns:saml="`+samlNS+`"><saml:NameID>user@example.com</saml:NameID></root>`)
 
-	key := generateRSAKey(t)
-	target := doc.DocumentElement().FirstChild().(*helium.Element)
+		key := generateRSAKey(t)
+		target := doc.DocumentElement().FirstChild().(*helium.Element)
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES256GCM).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
 
-	edElem, err := encryptor.EncryptElement(t.Context(), target)
-	require.NoError(t, err)
-	require.NotNil(t, edElem)
+		edElem, err := encryptor.EncryptElement(t.Context(), target)
+		require.NoError(t, err)
+		require.NotNil(t, edElem)
 
-	// EncryptedData now sits where saml:NameID was, so saml: is in-scope via
-	// its parent <root>.
-	require.NotNil(t, edElem.Parent())
+		// EncryptedData now sits where saml:NameID was, so saml: is in-scope via
+		// its parent <root>.
+		require.NotNil(t, edElem.Parent())
 
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "NameID", elem.LocalName())
-	require.Equal(t, samlNS, elem.URI(),
-		"saml: prefix declared on the ancestor must resolve")
-}
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "NameID", elem.LocalName())
+		require.Equal(t, samlNS, elem.URI(),
+			"saml: prefix declared on the ancestor must resolve")
+	})
 
-// D-ENC-006: decrypting an ELEMENT for a DETACHED EncryptedData (no parent)
-// whose own default xmlns is the XML-Encryption namespace must NOT let that
-// declaration shadow the decrypted element. An unprefixed plaintext element
-// must resolve to NO namespace.
-func TestDecryptElement_DetachedNoShadowing(t *testing.T) {
-	sessionKey := make([]byte, 32)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+	// D-ENC-006: decrypting an ELEMENT for a DETACHED EncryptedData (no parent)
+	// whose own default xmlns is the XML-Encryption namespace must NOT let that
+	// declaration shadow the decrypted element. An unprefixed plaintext element
+	// must resolve to NO namespace.
+	t.Run("element detached no shadowing", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
 
-	algorithm := xmlenc1.AES256GCM
+		algorithm := xmlenc1.AES256GCM
 
-	plaintext := []byte(`<child>v</child>`)
-	cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
-	require.NoError(t, err)
+		plaintext := []byte(`<child>v</child>`)
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
+		require.NoError(t, err)
 
-	doc := mustParseXML(t, `<root/>`)
+		doc := mustParseXML(t, `<root/>`)
 
-	ed := &xmlenc1.EncryptedData{
-		Type:             xmlenc1.TypeElement,
-		EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
-		CipherValue:      cipher,
-	}
-	edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
-	require.NoError(t, err)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeElement,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
-	require.Nil(t, edElem.Parent(), "EncryptedData must be detached for this test")
+		require.NoError(t, edElem.DeclareNamespace("", xmlenc1.NamespaceXMLEnc))
+		require.Nil(t, edElem.Parent(), "EncryptedData must be detached for this test")
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "child", elem.LocalName())
-	require.Equal(t, "", elem.URI(),
-		"detached EncryptedData's own xmlns must not shadow the decrypted element")
-}
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "child", elem.LocalName())
+		require.Equal(t, "", elem.URI(),
+			"detached EncryptedData's own xmlns must not shadow the decrypted element")
+	})
 
-// D-ENC-003: a payload prefix declared on EncryptedData's parent ancestor
-// still resolves when the content is parsed in the parent's context.
-func TestDecryptContent_PrefixOnParentAncestor(t *testing.T) {
-	sessionKey := make([]byte, 32)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+	// D-ENC-003: a payload prefix declared on EncryptedData's parent ancestor
+	// still resolves when the content is parsed in the parent's context.
+	t.Run("content prefix on parent ancestor", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
 
-	algorithm := xmlenc1.AES256GCM
+		algorithm := xmlenc1.AES256GCM
 
-	plaintext := []byte(`<app:item>v</app:item>`)
-	cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
-	require.NoError(t, err)
+		plaintext := []byte(`<app:item>v</app:item>`)
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, plaintext)
+		require.NoError(t, err)
 
-	const appNS = "http://example.com/app"
-	doc := mustParseXML(t,
-		`<root xmlns:app="`+appNS+`"><branch><wrapper/></branch></root>`)
+		const appNS = "http://example.com/app"
+		doc := mustParseXML(t,
+			`<root xmlns:app="`+appNS+`"><branch><wrapper/></branch></root>`)
 
-	ed := &xmlenc1.EncryptedData{
-		Type:             xmlenc1.TypeContent,
-		EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
-		CipherValue:      cipher,
-	}
-	edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
-	require.NoError(t, err)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeContent,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		edElem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	wrapper := doc.DocumentElement().FirstChild().FirstChild().(*helium.Element)
-	require.NoError(t, wrapper.AddChild(edElem))
+		wrapper := doc.DocumentElement().FirstChild().FirstChild().(*helium.Element)
+		require.NoError(t, wrapper.AddChild(edElem))
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	elem, ok := nodes[0].(*helium.Element)
-	require.True(t, ok, "decrypted node should be an element")
-	require.Equal(t, "item", elem.LocalName())
-	require.Equal(t, appNS, elem.URI(),
-		"app: prefix declared on parent ancestor must resolve")
+		elem, ok := nodes[0].(*helium.Element)
+		require.True(t, ok, "decrypted node should be an element")
+		require.Equal(t, "item", elem.LocalName())
+		require.Equal(t, appNS, elem.URI(),
+			"app: prefix declared on parent ancestor must resolve")
+	})
 }
