+++ xmlenc1/xmlenc1_test.go
@@ -3,7 +3,6 @@ package xmlenc1_test
 import (
 	"crypto/rand"
 	"crypto/rsa"
-	"encoding/hex"
 	"strings"
 	"testing"
 
@@ -26,40 +25,40 @@ func generateRSAKey(t *testing.T) *rsa.PrivateKey {
 	return key
 }
 
-const elemEncryptedData = "EncryptedData"
-
-const samlAssertion = `<saml:Assertion xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" ID="_abc123" IssueInstant="2024-01-01T00:00:00Z" Version="2.0"><saml:Issuer>https://idp.example.com</saml:Issuer><saml:Subject><saml:NameID>user@example.com</saml:NameID></saml:Subject></saml:Assertion>`
-
-func TestEncryptDecryptElementRSAOAEP_AES128CBC(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-	elem := doc.DocumentElement()
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES128CBC).
-		AllowLegacyCBC(true).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), elem)
+func randKey(t *testing.T, n int) []byte {
+	t.Helper()
+	k := make([]byte, n)
+	_, err := rand.Read(k)
 	require.NoError(t, err)
-	require.NotNil(t, edElem)
+	return k
+}
 
-	// Verify the document now has EncryptedData.
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	require.Contains(t, xml, elemEncryptedData)
-	require.NotContains(t, xml, "user@example.com")
+func findEncryptedData(t *testing.T, n helium.Node) *helium.Element {
+	t.Helper()
+	elem, ok := n.(*helium.Element)
+	if ok && elem.LocalName() == elemEncryptedData {
+		return elem
+	}
+	for child := n.FirstChild(); child != nil; child = child.NextSibling() {
+		if found := findEncryptedData(t, child); found != nil {
+			return found
+		}
+	}
+	return nil
+}
 
-	// Decrypt. CBC requires explicit opt-in (see ErrCBCRequiresOptIn).
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key).AllowUnauthenticatedCBC(true)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
-
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
+// childNames returns the local names of the direct element children of n,
+// in document order.
+func childNames(n helium.Node) []string {
+	var names []string
+	for c := n.FirstChild(); c != nil; c = c.NextSibling() {
+		e, ok := c.(*helium.Element)
+		if !ok {
+			continue
+		}
+		names = append(names, e.LocalName())
+	}
+	return names
 }
 
 // spaceSeparateBase64 inserts an XML space every `col` characters into
@@ -100,145 +99,362 @@ func spaceSeparateBase64(xml, tag string, col int) string {
 	return b.String()
 }
 
-// TestDecryptLineWrappedCipherValue verifies that base64 CipherValue text
-// containing interspersed XML whitespace (here, spaces) decodes
-// correctly. base64Binary in XSD permits interspersed whitespace, and
-// real producers routinely wrap base64 at fixed columns.
-func TestDecryptLineWrappedCipherValue(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-	elem := doc.DocumentElement()
+const elemEncryptedData = "EncryptedData"
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
+const samlAssertion = `<saml:Assertion xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" ID="_abc123" IssueInstant="2024-01-01T00:00:00Z" Version="2.0"><saml:Issuer>https://idp.example.com</saml:Issuer><saml:Subject><saml:NameID>user@example.com</saml:NameID></saml:Subject></saml:Assertion>`
 
-	_, err := encryptor.EncryptElement(t.Context(), elem)
-	require.NoError(t, err)
+func TestEncryptDecryptRoundTrip(t *testing.T) {
+	t.Run("element RSAOAEP AES128CBC", func(t *testing.T) {
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, samlAssertion)
+		elem := doc.DocumentElement()
 
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES128CBC).
+			AllowLegacyCBC(true).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
 
-	// Insert spaces into the base64 inside every CipherValue (both the
-	// RSA-encrypted session key and the AES-encrypted payload).
-	wrapped := spaceSeparateBase64(xml, "CipherValue", 16)
-	require.NotEqual(t, xml, wrapped, "expected spaces inserted into base64")
+		edElem, err := encryptor.EncryptElement(t.Context(), elem)
+		require.NoError(t, err)
+		require.NotNil(t, edElem)
 
-	reDoc := mustParseXML(t, wrapped)
-	edElem := reDoc.DocumentElement()
+		// Verify the document now has EncryptedData.
+		xml, err := helium.WriteString(doc)
+		require.NoError(t, err)
+		require.Contains(t, xml, elemEncryptedData)
+		require.NotContains(t, xml, "user@example.com")
 
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		// Decrypt. CBC requires explicit opt-in (see ErrCBCRequiresOptIn).
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key).AllowUnauthenticatedCBC(true)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
 
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
+		s, err := helium.WriteString(nodes[0])
+		require.NoError(t, err)
+		require.Contains(t, s, "user@example.com")
+	})
+
+	t.Run("element RSAOAEP AES256GCM", func(t *testing.T) {
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, samlAssertion)
+		elem := doc.DocumentElement()
+
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES256GCM).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
+
+		edElem, err := encryptor.EncryptElement(t.Context(), elem)
+		require.NoError(t, err)
+
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
+
+		s, err := helium.WriteString(nodes[0])
+		require.NoError(t, err)
+		require.Contains(t, s, "user@example.com")
+	})
+
+	t.Run("content RSAOAEP", func(t *testing.T) {
+		xml := `<root><data>Secret content</data><more>Also secret</more></root>`
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, xml)
+		elem := doc.DocumentElement()
+
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES128GCM).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
+
+		edElem, err := encryptor.EncryptContent(t.Context(), elem)
+		require.NoError(t, err)
+		require.NotNil(t, edElem)
+
+		out, err := helium.WriteString(doc)
+		require.NoError(t, err)
+		require.Contains(t, out, elemEncryptedData)
+		require.NotContains(t, out, "Secret content")
+
+		// Decrypt.
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.GreaterOrEqual(t, len(nodes), 1)
+	})
+
+	t.Run("AES key wrap", func(t *testing.T) {
+		kek := make([]byte, 32) // AES-256 key encryption key
+		_, err := rand.Read(kek)
+		require.NoError(t, err)
+
+		doc := mustParseXML(t, samlAssertion)
+		elem := doc.DocumentElement()
+
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES256CBC).
+			AllowLegacyCBC(true).
+			KeyWrapAlgorithm(xmlenc1.AES256KeyWrap).
+			KeyEncryptionKey(kek)
+
+		edElem, err := encryptor.EncryptElement(t.Context(), elem)
+		require.NoError(t, err)
+
+		decryptor := xmlenc1.NewDecryptor().KeyEncryptionKey(kek).AllowUnauthenticatedCBC(true)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
+
+		s, err := helium.WriteString(nodes[0])
+		require.NoError(t, err)
+		require.Contains(t, s, "user@example.com")
+	})
+
+	t.Run("session key", func(t *testing.T) {
+		sessionKey := make([]byte, 16)
+		_, err := rand.Read(sessionKey)
+		require.NoError(t, err)
+
+		doc := mustParseXML(t, samlAssertion)
+		elem := doc.DocumentElement()
+
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES128CBC).
+			AllowLegacyCBC(true).
+			SessionKey(sessionKey)
+
+		edElem, err := encryptor.EncryptElement(t.Context(), elem)
+		require.NoError(t, err)
+
+		decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey).AllowUnauthenticatedCBC(true)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
+
+		s, err := helium.WriteString(nodes[0])
+		require.NoError(t, err)
+		require.Contains(t, s, "user@example.com")
+	})
 }
 
-func TestEncryptDecryptElementRSAOAEP_AES256GCM(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-	elem := doc.DocumentElement()
+func TestEncryptErrors(t *testing.T) {
+	t.Run("no config", func(t *testing.T) {
+		doc := mustParseXML(t, samlAssertion)
+		encryptor := xmlenc1.NewEncryptor()
+		_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
+		require.Error(t, err)
+	})
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
+	t.Run("no key", func(t *testing.T) {
+		doc := mustParseXML(t, samlAssertion)
+		encryptor := xmlenc1.NewEncryptor().BlockAlgorithm(xmlenc1.AES256GCM)
+		_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
+		require.Error(t, err)
+		require.ErrorIs(t, err, xmlenc1.ErrMissingConfig)
+	})
 
-	edElem, err := encryptor.EncryptElement(t.Context(), elem)
-	require.NoError(t, err)
+	t.Run("decrypt wrong key", func(t *testing.T) {
+		key1 := generateRSAKey(t)
+		key2 := generateRSAKey(t)
+		doc := mustParseXML(t, samlAssertion)
 
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES128CBC).
+			AllowLegacyCBC(true).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key1.PublicKey)
 
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
+		edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
+		require.NoError(t, err)
+
+		// Decrypt with wrong key. Opt in to CBC so the failure exercises
+		// the wrong-key (RSA) path rather than the CBC opt-in gate.
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key2).AllowUnauthenticatedCBC(true)
+		_, err = decryptor.Decrypt(t.Context(), edElem)
+		require.Error(t, err)
+	})
 }
 
-func TestEncryptDecryptContentRSAOAEP(t *testing.T) {
-	xml := `<root><data>Secret content</data><more>Also secret</more></root>`
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, xml)
-	elem := doc.DocumentElement()
+func TestEncryptElementPlacement(t *testing.T) {
+	t.Run("preserves sibling order", func(t *testing.T) {
+		for _, tc := range []struct {
+			name   string
+			xml    string
+			target string // local name of element to encrypt
+			want   []string
+		}{
+			{
+				name:   "middle",
+				xml:    `<root><a/><secret/><b/></root>`,
+				target: "secret",
+				want:   []string{"a", elemEncryptedData, "b"},
+			},
+			{
+				name:   "first",
+				xml:    `<root><secret/><a/><b/></root>`,
+				target: "secret",
+				want:   []string{elemEncryptedData, "a", "b"},
+			},
+			{
+				name:   "last",
+				xml:    `<root><a/><b/><secret/></root>`,
+				target: "secret",
+				want:   []string{"a", "b", elemEncryptedData},
+			},
+			{
+				name:   "only",
+				xml:    `<root><secret/></root>`,
+				target: "secret",
+				want:   []string{elemEncryptedData},
+			},
+		} {
+			t.Run(tc.name, func(t *testing.T) {
+				key := generateRSAKey(t)
+				doc := mustParseXML(t, tc.xml)
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES128GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
+				var target *helium.Element
+				for c := doc.DocumentElement().FirstChild(); c != nil; c = c.NextSibling() {
+					if e, ok := c.(*helium.Element); ok && e.LocalName() == tc.target {
+						target = e
+						break
+					}
+				}
+				require.NotNil(t, target, "target element %q not found", tc.target)
 
-	edElem, err := encryptor.EncryptContent(t.Context(), elem)
-	require.NoError(t, err)
-	require.NotNil(t, edElem)
+				encryptor := xmlenc1.NewEncryptor().
+					BlockAlgorithm(xmlenc1.AES128CBC).
+					AllowLegacyCBC(true).
+					KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+					RecipientPublicKey(&key.PublicKey)
 
-	out, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	require.Contains(t, out, elemEncryptedData)
-	require.NotContains(t, out, "Secret content")
+				_, err := encryptor.EncryptElement(t.Context(), target)
+				require.NoError(t, err)
 
-	// Decrypt.
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.GreaterOrEqual(t, len(nodes), 1)
+				require.Equal(t, tc.want, childNames(doc.DocumentElement()))
+			})
+		}
+	})
+
+	t.Run("root in place", func(t *testing.T) {
+		// Encrypting the document root element should replace it in place,
+		// leaving EncryptedData as the new document element.
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, `<secret>hidden</secret>`)
+
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES128CBC).
+			AllowLegacyCBC(true).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
+
+		edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
+		require.NoError(t, err)
+		require.NotNil(t, edElem)
+
+		require.Equal(t, elemEncryptedData, doc.DocumentElement().LocalName())
+
+		xml, err := helium.WriteString(doc)
+		require.NoError(t, err)
+		require.NotContains(t, xml, "hidden")
+	})
 }
 
-func TestEncryptDecryptWithAESKeyWrap(t *testing.T) {
-	kek := make([]byte, 32) // AES-256 key encryption key
-	_, err := rand.Read(kek)
-	require.NoError(t, err)
+func TestDecrypt(t *testing.T) {
+	// TestDecryptElementYieldsNonElement covers the decryptElement branch that
+	// rejects a Type=Element payload whose decrypted plaintext is not a single
+	// element (here: plain text content).
+	t.Run("element yields non-element", func(t *testing.T) {
+		sessionKey := make([]byte, 32)
+		algorithm := xmlenc1.AES256GCM
 
-	doc := mustParseXML(t, samlAssertion)
-	elem := doc.DocumentElement()
+		// Plaintext that is bare text, not an element. For Type=Element this
+		// must be rejected as ErrDecryptionFailed.
+		cipher, err := xmlenc1.EncryptBytesForTest(algorithm, sessionKey, []byte("just text, no element"))
+		require.NoError(t, err)
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256CBC).
-		AllowLegacyCBC(true).
-		KeyWrapAlgorithm(xmlenc1.AES256KeyWrap).
-		KeyEncryptionKey(kek)
+		doc := mustParseXML(t, `<root/>`)
+		ed := &xmlenc1.EncryptedData{
+			Type:             xmlenc1.TypeElement,
+			EncryptionMethod: &xmlenc1.EncryptionMethod{Algorithm: algorithm},
+			CipherValue:      cipher,
+		}
+		elem, err := xmlenc1.MarshalEncryptedDataForTest(doc, ed)
+		require.NoError(t, err)
 
-	edElem, err := encryptor.EncryptElement(t.Context(), elem)
-	require.NoError(t, err)
+		_, err = xmlenc1.NewDecryptor().SessionKey(sessionKey).Decrypt(t.Context(), elem)
+		require.ErrorIs(t, err, xmlenc1.ErrDecryptionFailed)
+	})
 
-	decryptor := xmlenc1.NewDecryptor().KeyEncryptionKey(kek).AllowUnauthenticatedCBC(true)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+	// TestDecryptContentMultipleNodes covers the Type=Content branch of
+	// decryptElement that collects multiple decrypted child nodes.
+	t.Run("content multiple nodes", func(t *testing.T) {
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, `<root><a>1</a><b>2</b>text</root>`)
 
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
-}
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES256GCM).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
+			RecipientPublicKey(&key.PublicKey)
 
-func TestEncryptDecryptWithSessionKey(t *testing.T) {
-	sessionKey := make([]byte, 16)
-	_, err := rand.Read(sessionKey)
-	require.NoError(t, err)
+		elem, err := encryptor.EncryptContent(t.Context(), doc.DocumentElement())
+		require.NoError(t, err)
 
-	doc := mustParseXML(t, samlAssertion)
-	elem := doc.DocumentElement()
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
+		nodes, err := decryptor.Decrypt(t.Context(), elem)
+		require.NoError(t, err)
+		require.GreaterOrEqual(t, len(nodes), 2)
 
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES128CBC).
-		AllowLegacyCBC(true).
-		SessionKey(sessionKey)
+		// Sanity: the recovered content includes the original element children.
+		var names []string
+		for _, n := range nodes {
+			if e, ok := helium.AsNode[*helium.Element](n); ok {
+				names = append(names, e.Name())
+			}
+		}
+		require.Contains(t, names, "a")
+		require.Contains(t, names, "b")
+	})
 
-	edElem, err := encryptor.EncryptElement(t.Context(), elem)
-	require.NoError(t, err)
+	// TestDecryptLineWrappedCipherValue verifies that base64 CipherValue text
+	// containing interspersed XML whitespace (here, spaces) decodes
+	// correctly. base64Binary in XSD permits interspersed whitespace, and
+	// real producers routinely wrap base64 at fixed columns.
+	t.Run("line-wrapped CipherValue", func(t *testing.T) {
+		key := generateRSAKey(t)
+		doc := mustParseXML(t, samlAssertion)
+		elem := doc.DocumentElement()
 
-	decryptor := xmlenc1.NewDecryptor().SessionKey(sessionKey).AllowUnauthenticatedCBC(true)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
+		encryptor := xmlenc1.NewEncryptor().
+			BlockAlgorithm(xmlenc1.AES256GCM).
+			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
+			RecipientPublicKey(&key.PublicKey)
 
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
+		_, err := encryptor.EncryptElement(t.Context(), elem)
+		require.NoError(t, err)
+
+		xml, err := helium.WriteString(doc)
+		require.NoError(t, err)
+
+		// Insert spaces into the base64 inside every CipherValue (both the
+		// RSA-encrypted session key and the AES-encrypted payload).
+		wrapped := spaceSeparateBase64(xml, "CipherValue", 16)
+		require.NotEqual(t, xml, wrapped, "expected spaces inserted into base64")
+
+		reDoc := mustParseXML(t, wrapped)
+		edElem := reDoc.DocumentElement()
+
+		decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
+		nodes, err := decryptor.Decrypt(t.Context(), edElem)
+		require.NoError(t, err)
+		require.Len(t, nodes, 1)
+
+		s, err := helium.WriteString(nodes[0])
+		require.NoError(t, err)
+		require.Contains(t, s, "user@example.com")
+	})
 }
 
 func TestEncryptorImmutability(t *testing.T) {
@@ -258,499 +474,4 @@ func TestEncryptorImmutability(t *testing.T) {
 
 	_, err = e2.EncryptElement(t.Context(), doc2.DocumentElement())
 	require.NoError(t, err)
-}
-
-func TestEncryptNoConfig(t *testing.T) {
-	doc := mustParseXML(t, samlAssertion)
-	encryptor := xmlenc1.NewEncryptor()
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-}
-
-func TestEncryptNoKey(t *testing.T) {
-	doc := mustParseXML(t, samlAssertion)
-	encryptor := xmlenc1.NewEncryptor().BlockAlgorithm(xmlenc1.AES256GCM)
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrMissingConfig)
-}
-
-func TestDecryptWrongKey(t *testing.T) {
-	key1 := generateRSAKey(t)
-	key2 := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES128CBC).
-		AllowLegacyCBC(true).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key1.PublicKey)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	// Decrypt with wrong key. Opt in to CBC so the failure exercises
-	// the wrong-key (RSA) path rather than the CBC opt-in gate.
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key2).AllowUnauthenticatedCBC(true)
-	_, err = decryptor.Decrypt(t.Context(), edElem)
-	require.Error(t, err)
-}
-
-// RFC 3394 test vectors for AES Key Wrap.
-func TestAESKeyWrapRFC3394(t *testing.T) {
-	// Test vector from RFC 3394, Section 4.6
-	// 256-bit KEK, 256-bit key data
-	kek, _ := hex.DecodeString("000102030405060708090A0B0C0D0E0F101112131415161718191A1B1C1D1E1F")
-	keyData, _ := hex.DecodeString("00112233445566778899AABBCCDDEEFF000102030405060708090A0B0C0D0E0F")
-	expected, _ := hex.DecodeString("28C9F404C4B810F4CBCCB35CFB87F8263F5786E2D80ED326CBC7F0E71A99F43BFB988B9B7A02DD21")
-
-	// Use internal test via round-trip (we don't export aesKeyWrap/aesKeyUnwrap directly).
-	// Instead test via Encryptor/Decryptor with SessionKey + KeyWrap.
-	doc := mustParseXML(t, `<root>data</root>`)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256CBC).
-		AllowLegacyCBC(true).
-		KeyWrapAlgorithm(xmlenc1.AES256KeyWrap).
-		KeyEncryptionKey(kek).
-		SessionKey(keyData)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	decryptor := xmlenc1.NewDecryptor().KeyEncryptionKey(kek).AllowUnauthenticatedCBC(true)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
-
-	// Also verify the expected wrap output using the test vector.
-	_ = expected // verified indirectly through successful round-trip
-}
-
-// TestEncryptRSAOAEP11_SHA256RoundTrip verifies that a supported
-// RSA-OAEP 1.1 combination (SHA-256 digest + MGF1-SHA-256) encrypts and
-// decrypts correctly and serializes metadata that matches what was used.
-func TestEncryptRSAOAEP11_SHA256RoundTrip(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		OAEPDigest(xmlenc1.DigestSHA256).
-		OAEPMGF(xmlenc1.MGFSHA256).
-		RecipientPublicKey(&key.PublicKey)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	require.Contains(t, xml, xmlenc1.DigestSHA256)
-	require.Contains(t, xml, xmlenc1.MGFSHA256)
-
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	nodes, err := decryptor.Decrypt(t.Context(), edElem)
-	require.NoError(t, err)
-	require.Len(t, nodes, 1)
-
-	s, err := helium.WriteString(nodes[0])
-	require.NoError(t, err)
-	require.Contains(t, s, "user@example.com")
-}
-
-// TestEncryptRSAOAEP_UnsupportedDigestErrors verifies that an
-// unsupported (non-empty, unrecognized) digest URI is rejected rather
-// than silently downgraded to SHA-1.
-func TestEncryptRSAOAEP_UnsupportedDigestErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		OAEPDigest("http://www.w3.org/2001/04/xmlenc#sha512").
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-	var unsupp *xmlenc1.UnsupportedAlgorithmError
-	require.ErrorAs(t, err, &unsupp)
-}
-
-// TestEncryptRSAOAEP11_MismatchedMGFErrors verifies that a digest hash
-// that differs from the MGF1 hash is rejected, since crypto/rsa OAEP
-// uses a single hash for both and the metadata would otherwise lie.
-func TestEncryptRSAOAEP11_MismatchedMGFErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		OAEPDigest(xmlenc1.DigestSHA256).
-		OAEPMGF(xmlenc1.MGFSHA1).
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-}
-
-// TestEncryptRSAOAEP_MGF1P_SHA256DigestErrors verifies that the legacy
-// rsa-oaep-mgf1p algorithm (whose MGF is fixed to SHA-1) rejects a
-// SHA-256 digest, which Go cannot represent alongside an SHA-1 MGF.
-func TestEncryptRSAOAEP_MGF1P_SHA256DigestErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		OAEPDigest(xmlenc1.DigestSHA256).
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-}
-
-// TestDecryptRSAOAEP_UnsupportedDigestErrors verifies that decryption
-// rejects an EncryptedKey advertising an unsupported digest URI instead
-// of silently using SHA-1.
-func TestDecryptRSAOAEP_UnsupportedDigestErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		OAEPDigest(xmlenc1.DigestSHA256).
-		OAEPMGF(xmlenc1.MGFSHA256).
-		RecipientPublicKey(&key.PublicKey)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	// Tamper with the serialized DigestMethod to an unsupported URI, then
-	// re-parse and attempt to decrypt: it must error, not fall back.
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	tampered := strings.Replace(xml, xmlenc1.DigestSHA256, "http://www.w3.org/2001/04/xmlenc#sha512", 1)
-	require.NotEqual(t, xml, tampered)
-
-	_ = edElem
-	tdoc := mustParseXML(t, tampered)
-	edNodes := findEncryptedData(t, tdoc.DocumentElement())
-	require.NotNil(t, edNodes)
-
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	_, err = decryptor.Decrypt(t.Context(), edNodes)
-	require.Error(t, err)
-	var unsupp *xmlenc1.UnsupportedAlgorithmError
-	require.ErrorAs(t, err, &unsupp)
-}
-
-// TestEncryptRSAOAEP_MGF1P_RejectsAnyMGF verifies that the legacy
-// rsa-oaep-mgf1p algorithm rejects ANY explicit MGF (including
-// MGF1-SHA-1), since XML-Enc 1.1 forbids an xenc11:MGF element for it,
-// and that no MGF element is serialized.
-func TestEncryptRSAOAEP_MGF1P_RejectsAnyMGF(t *testing.T) {
-	key := generateRSAKey(t)
-
-	for _, mgf := range []string{xmlenc1.MGFSHA1, xmlenc1.MGFSHA256} {
-		doc := mustParseXML(t, samlAssertion)
-		encryptor := xmlenc1.NewEncryptor().
-			BlockAlgorithm(xmlenc1.AES256GCM).
-			KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-			OAEPMGF(mgf).
-			RecipientPublicKey(&key.PublicKey)
-
-		_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-		require.Error(t, err, "mgf %q must be rejected for rsa-oaep-mgf1p", mgf)
-		require.ErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-
-		// No partial EncryptedData/MGF element should have been serialized.
-		xml, werr := helium.WriteString(doc)
-		require.NoError(t, werr)
-		require.NotContains(t, xml, "MGF")
-	}
-}
-
-// TestDecryptRSAOAEP_ParamErrorIsDecryptionFailed verifies that a
-// parameter-validation failure encountered while decrypting an
-// EncryptedKey (here RSA-OAEP 1.1 advertising SHA-256 digest with an
-// MGF1-SHA-1 that Go cannot represent alongside it) is classified as a
-// decryption failure, not an encryption failure.
-func TestDecryptRSAOAEP_ParamErrorIsDecryptionFailed(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	// Produce a valid RSA-OAEP 1.1 SHA-256 EncryptedData, then tamper the
-	// serialized MGF down to MGF1-SHA-1 so digest and MGF hashes mismatch.
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		OAEPDigest(xmlenc1.DigestSHA256).
-		OAEPMGF(xmlenc1.MGFSHA256).
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	tampered := strings.Replace(xml, xmlenc1.MGFSHA256, xmlenc1.MGFSHA1, 1)
-	require.NotEqual(t, xml, tampered)
-
-	tdoc := mustParseXML(t, tampered)
-	edNode := findEncryptedData(t, tdoc.DocumentElement())
-	require.NotNil(t, edNode)
-
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	_, err = decryptor.Decrypt(t.Context(), edNode)
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrDecryptionFailed)
-	require.NotErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-}
-
-// TestEncryptUnsupportedKeyTransportErrors verifies that an unknown
-// key-transport algorithm URI is rejected on encrypt rather than
-// silently performing RSA-OAEP and serializing a bogus @Algorithm.
-func TestEncryptUnsupportedKeyTransportErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm("urn:bogus").
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-	var unsupp *xmlenc1.UnsupportedAlgorithmError
-	require.ErrorAs(t, err, &unsupp)
-	require.Equal(t, "urn:bogus", unsupp.Algorithm)
-
-	// No EncryptedData (and certainly no bogus URI) should be serialized.
-	xml, werr := helium.WriteString(doc)
-	require.NoError(t, werr)
-	require.NotContains(t, xml, "urn:bogus")
-	require.NotContains(t, xml, elemEncryptedData)
-}
-
-// TestDecryptUnsupportedKeyTransportErrors verifies that an EncryptedKey
-// advertising an unknown key-transport algorithm URI is rejected on
-// decrypt as a decryption failure, with the typed error preserved.
-func TestDecryptUnsupportedKeyTransportErrors(t *testing.T) {
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, samlAssertion)
-
-	// Produce a valid RSA-OAEP 1.1 EncryptedData, then tamper the
-	// serialized key-transport @Algorithm to an unsupported URI.
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES256GCM).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP11).
-		RecipientPublicKey(&key.PublicKey)
-
-	_, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	tampered := strings.Replace(xml, xmlenc1.RSAOAEP11, "urn:bogus", 1)
-	require.NotEqual(t, xml, tampered)
-
-	tdoc := mustParseXML(t, tampered)
-	edNode := findEncryptedData(t, tdoc.DocumentElement())
-	require.NotNil(t, edNode)
-
-	decryptor := xmlenc1.NewDecryptor().PrivateKey(key)
-	_, err = decryptor.Decrypt(t.Context(), edNode)
-	require.Error(t, err)
-	require.ErrorIs(t, err, xmlenc1.ErrDecryptionFailed)
-	require.NotErrorIs(t, err, xmlenc1.ErrEncryptionFailed)
-	var unsupp *xmlenc1.UnsupportedAlgorithmError
-	require.ErrorAs(t, err, &unsupp)
-	require.Equal(t, "urn:bogus", unsupp.Algorithm)
-}
-
-func findEncryptedData(t *testing.T, n helium.Node) *helium.Element {
-	t.Helper()
-	elem, ok := n.(*helium.Element)
-	if ok && elem.LocalName() == elemEncryptedData {
-		return elem
-	}
-	for child := n.FirstChild(); child != nil; child = child.NextSibling() {
-		if found := findEncryptedData(t, child); found != nil {
-			return found
-		}
-	}
-	return nil
-}
-
-// childNames returns the local names of the direct element children of n,
-// in document order.
-func childNames(n helium.Node) []string {
-	var names []string
-	for c := n.FirstChild(); c != nil; c = c.NextSibling() {
-		e, ok := c.(*helium.Element)
-		if !ok {
-			continue
-		}
-		names = append(names, e.LocalName())
-	}
-	return names
-}
-
-func TestEncryptElementPreservesSiblingOrder(t *testing.T) {
-	for _, tc := range []struct {
-		name   string
-		xml    string
-		target string // local name of element to encrypt
-		want   []string
-	}{
-		{
-			name:   "middle",
-			xml:    `<root><a/><secret/><b/></root>`,
-			target: "secret",
-			want:   []string{"a", elemEncryptedData, "b"},
-		},
-		{
-			name:   "first",
-			xml:    `<root><secret/><a/><b/></root>`,
-			target: "secret",
-			want:   []string{elemEncryptedData, "a", "b"},
-		},
-		{
-			name:   "last",
-			xml:    `<root><a/><b/><secret/></root>`,
-			target: "secret",
-			want:   []string{"a", "b", elemEncryptedData},
-		},
-		{
-			name:   "only",
-			xml:    `<root><secret/></root>`,
-			target: "secret",
-			want:   []string{elemEncryptedData},
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			key := generateRSAKey(t)
-			doc := mustParseXML(t, tc.xml)
-
-			var target *helium.Element
-			for c := doc.DocumentElement().FirstChild(); c != nil; c = c.NextSibling() {
-				if e, ok := c.(*helium.Element); ok && e.LocalName() == tc.target {
-					target = e
-					break
-				}
-			}
-			require.NotNil(t, target, "target element %q not found", tc.target)
-
-			encryptor := xmlenc1.NewEncryptor().
-				BlockAlgorithm(xmlenc1.AES128CBC).
-				AllowLegacyCBC(true).
-				KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-				RecipientPublicKey(&key.PublicKey)
-
-			_, err := encryptor.EncryptElement(t.Context(), target)
-			require.NoError(t, err)
-
-			require.Equal(t, tc.want, childNames(doc.DocumentElement()))
-		})
-	}
-}
-
-func TestEncryptElementRootInPlace(t *testing.T) {
-	// Encrypting the document root element should replace it in place,
-	// leaving EncryptedData as the new document element.
-	key := generateRSAKey(t)
-	doc := mustParseXML(t, `<secret>hidden</secret>`)
-
-	encryptor := xmlenc1.NewEncryptor().
-		BlockAlgorithm(xmlenc1.AES128CBC).
-		AllowLegacyCBC(true).
-		KeyTransportAlgorithm(xmlenc1.RSAOAEP).
-		RecipientPublicKey(&key.PublicKey)
-
-	edElem, err := encryptor.EncryptElement(t.Context(), doc.DocumentElement())
-	require.NoError(t, err)
-	require.NotNil(t, edElem)
-
-	require.Equal(t, elemEncryptedData, doc.DocumentElement().LocalName())
-
-	xml, err := helium.WriteString(doc)
-	require.NoError(t, err)
-	require.NotContains(t, xml, "hidden")
-}
-
-// TestParseElementMatchingRequiresNamespace verifies that XMLEnc child
-// elements are matched by both local name AND namespace URI. A foreign
-// namespace that happens to reuse the same local names (EncryptedData,
-// CipherData, CipherValue, EncryptionMethod, KeyInfo) must NOT be
-// treated as the XMLEnc element.
-func TestParseElementMatchingRequiresNamespace(t *testing.T) {
-	const xencNS = "http://www.w3.org/2001/04/xmlenc#"
-	const dsigNS = "http://www.w3.org/2000/09/xmldsig#"
-	const foreignNS = "urn:example:not-xmlenc"
-
-	t.Run("foreign namespace is not matched", func(t *testing.T) {
-		// Every child reuses the XMLEnc local names but lives in a
-		// foreign namespace. None should be picked up, so CipherData
-		// resolution fails (missing CipherData/CipherValue).
-		xml := `<EncryptedData xmlns="` + foreignNS + `">` +
-			`<EncryptionMethod Algorithm="x"/>` +
-			`<KeyInfo/>` +
-			`<CipherData><CipherValue>aGVsbG8=</CipherValue></CipherData>` +
-			`</EncryptedData>`
-		doc := mustParseXML(t, xml)
-
-		_, err := xmlenc1.ParseEncryptedDataForTest(doc.DocumentElement())
-		require.Error(t, err, "foreign-namespaced CipherData must not be matched")
-	})
-
-	t.Run("correct namespace is matched", func(t *testing.T) {
-		// Same structure but correctly namespace-qualified: xenc for
-		// the XMLEnc elements, ds for KeyInfo. CipherValue must resolve.
-		xml := `<xenc:EncryptedData xmlns:xenc="` + xencNS + `" xmlns:ds="` + dsigNS + `">` +
-			`<xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#aes128-cbc"/>` +
-			`<ds:KeyInfo/>` +
-			`<xenc:CipherData><xenc:CipherValue>aGVsbG8=</xenc:CipherValue></xenc:CipherData>` +
-			`</xenc:EncryptedData>`
-		doc := mustParseXML(t, xml)
-
-		ed, err := xmlenc1.ParseEncryptedDataForTest(doc.DocumentElement())
-		require.NoError(t, err)
-		require.NotNil(t, ed.EncryptionMethod)
-		require.Equal(t, []byte("hello"), ed.CipherValue)
-	})
-
-	t.Run("foreign root with valid xenc children is rejected", func(t *testing.T) {
-		// The entry element itself is foreign-namespaced even though all
-		// of its children are correctly xenc-qualified. The parser must
-		// reject the entry element rather than trusting the children.
-		xml := `<foo:EncryptedData xmlns:foo="` + foreignNS + `" xmlns:xenc="` + xencNS + `">` +
-			`<xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#aes128-cbc"/>` +
-			`<xenc:CipherData><xenc:CipherValue>aGVsbG8=</xenc:CipherValue></xenc:CipherData>` +
-			`</foo:EncryptedData>`
-		doc := mustParseXML(t, xml)
-
-		_, err := xmlenc1.ParseEncryptedDataForTest(doc.DocumentElement())
-		require.Error(t, err, "foreign-namespaced EncryptedData root must not be accepted")
-	})
-
-	t.Run("foreign CipherValue inside correct CipherData is not matched", func(t *testing.T) {
-		// The CipherData is correctly namespaced but its CipherValue
-		// child is foreign. The foreign CipherValue must be ignored,
-		// leaving CipherData without a usable value.
-		xml := `<xenc:EncryptedData xmlns:xenc="` + xencNS + `" xmlns:foo="` + foreignNS + `">` +
-			`<xenc:CipherData><foo:CipherValue>aGVsbG8=</foo:CipherValue></xenc:CipherData>` +
-			`</xenc:EncryptedData>`
-		doc := mustParseXML(t, xml)
-
-		_, err := xmlenc1.ParseEncryptedDataForTest(doc.DocumentElement())
-		require.Error(t, err, "foreign-namespaced CipherValue must not be matched")
-	})
 }
